Make TUI backend contracts explicit
Give the backend package and chat command entrypoint typed contracts so command wiring can launch TUI behavior without importing raw terminal implementation details.

Constraint: the backend runtime boundary must remain independent from prompt-toolkit presentation objects and raw ChatApplication ownership.

Rejected: keeping chat_cmd.py as the owner of TUI compatibility wiring | it makes Typer command import cost and terminal behavior evolve together.

Confidence: high

Scope-risk: moderate

Directive: keep future chat command extensions typed at the entrypoint and route terminal-specific compatibility through TUI modules.

Tested: pre-squash tree passed scoped TUI/CLI pytest, ruff, mypy, compileall, and import-smoke gates; post-squash tree identity checked before push.

Not-tested: manual interactive full-screen TUI session not run during history cleanup.
