+++ src/opensquilla/cli/tui/backend/__init__.py
@@ -1,0 +1,11 @@
+"""Pure backend contracts, state, and runtime for OpenSquilla TUI surfaces."""
+
+from __future__ import annotations
+
+__all__ = [
+    "contracts",
+    "events",
+    "output_binding",
+    "runtime",
+    "state",
+]
